+++ src/integrations/mod.rs
@@ -18,7 +18,7 @@ mod error;
 pub use error::VectorLoaderError;
 
 mod asset;
-pub use asset::{VelloAsset, VelloAssetAlignment};
+pub use asset::{VelloAsset, VelloAssetAnchor};
 
 #[derive(Clone)]
 pub enum VectorFile {