+++ src/integrations/svg/asset_loader.rs
@@ -39,13 +39,13 @@ impl AssetLoader for VelloSvgLoader {
             debug!("parsing {}...", load_context.path().display());
             match ext {
                 "svg" => {
-                    let vello_vector = load_svg_from_bytes(&bytes)?;
+                    let asset = load_svg_from_bytes(&bytes)?;
                     info!(
                         path = format!("{}", load_context.path().display()),
-                        size = format!("{:?}", (vello_vector.width, vello_vector.height)),
+                        size = format!("{:?}", (asset.width, asset.height)),
                         "finished parsing svg asset"
                     );
-                    Ok(vello_vector)
+                    Ok(asset)
                 }
                 ext => Err(VectorLoaderError::Io(std::io::Error::new(
                     std::io::ErrorKind::InvalidData,